Do no require a dependency that is not required
First of all, do not require dependencies that are not required for PHPMailer to work. It works perfectly without. Secondly, the majority of the users will not use Gmail's XOAUTH2 authentication system and do not need these additional dependencies. Thirdly, guzzle and league packages are used a lot. Therefore you make it hard for people that were already using these packages to newer versions of PHPMailer.